fix(php81): replace deprecated (double) cast with (float)
srand() requires an int; microtime() returns float.
Cast chain: (float)microtime() * 1000000 then (int) for srand.

+++ app/operators/include/management/pages_common.php
@@ -37,7 +37,7 @@ function createPassword($length, $chars) {
         $chars = "abcdefghijkmnpqrstuvwxyzABCDEFGHJKMNPQRSTUVWXYZ23456789";
     }
 
-    srand((double)microtime()*1000000);
+    srand((int)((float)microtime()*1000000));
     $i = 0;
     $pass = '';
